Reorganize AcceleratorPool and EvaluationResult
