+++ CMakeLists.txt
@@ -16,10 +16,11 @@ set(BINDINGS_SRC
     bindings/accelergy_interface.cpp
     bindings/bindings.cpp
     bindings/config.cpp
-    bindings/mapper.cpp
     bindings/mapping.cpp
     bindings/mapspace.cpp
+    bindings/model/accelerator-pool.cpp
     bindings/model/engine.cpp
+    bindings/model/eval-result.cpp
     bindings/model/sparse_optimization_info.cpp
     bindings/problem.cpp
     bindings/search.cpp)
