+++ bindings/bindings.cpp
@@ -3,7 +3,9 @@
 #include <variant>
 #include <vector>
 
+#include "bindings/model/accelerator-pool.h"
 #include "bindings/model/bindings.h"
+#include "bindings/model/eval-result.h"
 
 #define STRINGIFY(x) #x
 #define MACRO_STRINGIFY(x) STRINGIFY(x)
@@ -13,11 +15,12 @@ PYBIND11_MODULE(bindings, m) {
 
   BindAccelergyInterface(m);
   BindConfigClasses(m);
-  BindMapperClasses(m);
   BindMappingClasses(m);
   BindMapspaceClasses(m);
 
+  model_bindings::BindAcceleratorPool(m);
   model_bindings::BindEngine(m);
+  model_bindings::BindEvaluationResult(m);
   model_bindings::BindLevel(m);
   model_bindings::BindSparseOptimizationInfo(m);
   model_bindings::BindTopology(m);
